requirements.txt
# Core
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
yfinance==0.2.28
python-dotenv==1.0.1

# API
flask==3.0.0
flask-jwt-extended==4.5.3

# Dashboard
streamlit==1.28.0
plotly==5.18.0
streamlit-folium==0.14.0
folium==0.14.0
statsmodels==0.14.0


-e .